training definition contains estimated duration attribute.

Closes #268

See merge request kypo2/services-and-portlets/kypo2-training!240

+++ kypo2-persistence-training/src/main/java/cz/muni/ics/kypo/training/persistence/model/AbstractLevel.java
@@ -23,7 +23,7 @@ public abstract class AbstractLevel implements Serializable {
     @OneToOne(fetch = FetchType.LAZY, optional = true, cascade = CascadeType.ALL)
     private SnapshotHook snapshotHook;
     @Column(name = "estimated_duration")
-    private int estimatedDuration;
+    private long estimatedDuration;
     @Column(name = "order_in_training_definition", nullable = false)
     private int order;
     @ManyToOne(fetch = FetchType.LAZY)
@@ -65,11 +65,11 @@ public abstract class AbstractLevel implements Serializable {
         this.snapshotHook = snapshotHook;
     }
 
-    public int getEstimatedDuration() {
+    public long getEstimatedDuration() {
         return estimatedDuration;
     }
 
-    public void setEstimatedDuration(int estimatedDuration) {
+    public void setEstimatedDuration(long estimatedDuration) {
         this.estimatedDuration = estimatedDuration;
     }
